Fix delivery of retained messages with # wildcard.
The duplicate retained message fix broke retained messages on foo when
subscribing to foo/# This is because the behaviour before was incorrect and
retained messages were being stored on the "#" below the "foo". The duplicate
fix stopped this, so those messages were no longer stored as retained and so
were not delivered on subscription.